Move dbconnector
seperate dbconnector file fto the core folder
