+++ package.json
@@ -14,6 +14,7 @@
     "glue": "^4.2.1",
     "hapi": "^16.6.2",
     "inert": "^4.2.1",
+    "nodemon": "^1.12.4",
     "orientjs": "^2.2.7"
   }
 }
