+++ plugins/bot/bot-dialog.js
@@ -1,6 +1,7 @@
 var bot = require("./bot").bot;
 var recognizer = require("./bot").recognizer;
-var db = require("./dbconnector"); //OrintDb setup
+var db = require("../../core/db/dbconnector"); //OrintDb setup
+const builder = require('botbuilder');
 
 bot.recognizer(recognizer);
