+++ plugins/bot/bot.js
