--- plugins/bot/dbconnector.js
@@ -1,14 +1,0 @@
-var OrientDB = require('orientjs');
-var dbServer = OrientDB({
-    host: 'localhost',
-    port: 2424,
-    username: 'root',
-    password: 'root'
-});
-var db = dbServer.use({
-    name: 'tktnet',
-    username: 'root',
-    password: 'root'
-});
-
-module.exports = db;
