+++ plugins/bot/hapi-extensions.js
