+++ plugins/bot/index.js
