+++ plugins/bot/routes.js
