+++ plugins/client/index.js
