+++ plugins/client/routes.js
